Desktop/projects/Fitapp-fixierun-full/fitapp-project/python_backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
bidict==0.23.1
bitarray==3.3.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.1.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
cytoolz==1.0.1
dnspython==2.7.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
Flask==3.1.0
Flask-SocketIO==5.5.1
fonttools==4.57.0
frozenlist==1.6.0
geographiclib==2.0
geopy==2.4.1
h11==0.16.0
hexbytes==1.3.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
multidict==6.4.3
numpy==2.2.5
packaging==25.0
pandas==2.2.3
parsimonious==0.10.0
pillow==11.2.1
propcache==0.3.1
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.12.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.2
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
seaborn==0.13.2
simple-websocket==1.1.0
six==1.17.0
toolz==1.0.0
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
web3==7.10.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.0

# Dashboard dependencies
streamlit==1.36.0
plotly==5.20.0
pydeck==0.8.0
streamlit-authenticator==0.3.0
streamlit-folium==0.15.0
streamlit-extras==0.3.6
folium==0.14.0
watchdog==4.0.0
websocket-client==1.7.0
protobuf==4.25.3
leafmap==0.27.1
